+++ core/tauri/src/image/plugin.rs
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: MIT
 
 use crate::plugin::{Builder, TauriPlugin};
-use crate::{command, AppHandle, Image, Manager, ResourceId, Runtime};
+use crate::{command, image::Image, AppHandle, Manager, ResourceId, Runtime};
 
 #[command(root = "crate")]
 fn new<R: Runtime>(
@@ -31,36 +31,6 @@ fn from_bytes<R: Runtime>(app: AppHandle<R>, bytes: Vec<u8>) -> crate::Result<Re
 #[command(root = "crate")]
 fn from_bytes() -> std::result::Result<(), &'static str> {
   Err("from_bytes is only supported if the `image-ico` or `image-png` Cargo features are enabled")
-}
-
-#[cfg(feature = "image-ico")]
-#[command(root = "crate")]
-fn from_ico_bytes<R: Runtime>(app: AppHandle<R>, bytes: Vec<u8>) -> crate::Result<ResourceId> {
-  let image = Image::from_ico_bytes(&bytes)?.to_owned();
-  let mut resources_table = app.resources_table();
-  let rid = resources_table.add(image);
-  Ok(rid)
-}
-
-#[cfg(not(feature = "image-ico"))]
-#[command(root = "crate")]
-fn from_ico_bytes() -> std::result::Result<(), &'static str> {
-  Err("from_ico_bytes is only supported if the `image-ico` Cargo feature is enabled")
-}
-
-#[cfg(feature = "image-png")]
-#[command(root = "crate")]
-fn from_png_bytes<R: Runtime>(app: AppHandle<R>, bytes: Vec<u8>) -> crate::Result<ResourceId> {
-  let image = Image::from_png_bytes(&bytes)?.to_owned();
-  let mut resources_table = app.resources_table();
-  let rid = resources_table.add(image);
-  Ok(rid)
-}
-
-#[cfg(not(feature = "image-png"))]
-#[command(root = "crate")]
-fn from_png_bytes() -> std::result::Result<(), &'static str> {
-  Err("from_png_bytes is only supported if the `image-ico` Cargo feature is enabled")
 }
 
 #[cfg(any(feature = "image-ico", feature = "image-png"))]
@@ -103,14 +73,7 @@ fn height<R: Runtime>(app: AppHandle<R>, rid: ResourceId) -> crate::Result<u32> 
 pub fn init<R: Runtime>() -> TauriPlugin<R> {
   Builder::new("image")
     .invoke_handler(crate::generate_handler![
-      new,
-      from_bytes,
-      from_ico_bytes,
-      from_png_bytes,
-      from_path,
-      rgba,
-      width,
-      height
+      new, from_bytes, from_path, rgba, width, height
     ])
     .build()
 }