+++ tooling/api/src/image.ts
@@ -45,42 +45,6 @@ export class Image extends Resource {
   }
 
   /**
-   * Creates a new image using the provided png bytes.
-   *
-   * Note that you need the `image-png` Cargo features to use this API.
-   * To enable it, change your Cargo.toml file:
-   * ```toml
-   * [dependencies]
-   * tauri = { version = "...", features = ["...", "image-png"] }
-   * ```
-   */
-  static async fromPngBytes(
-    bytes: number[] | Uint8Array | ArrayBuffer
-  ): Promise<Image> {
-    return invoke<number>('plugin:image|from_png_bytes', {
-      bytes: transformImage(bytes)
-    }).then((rid) => new Image(rid))
-  }
-
-  /**
-   * Creates a new image using the provided ico bytes.
-   *
-   * Note that you need the `image-ico` Cargo features to use this API.
-   * To enable it, change your Cargo.toml file:
-   * ```toml
-   * [dependencies]
-   * tauri = { version = "...", features = ["...", "image-ico"] }
-   * ```
-   */
-  static async fromIcoBytes(
-    bytes: number[] | Uint8Array | ArrayBuffer
-  ): Promise<Image> {
-    return invoke<number>('plugin:image|from_ico_bytes', {
-      bytes: transformImage(bytes)
-    }).then((rid) => new Image(rid))
-  }
-
-  /**
    * Creates a new image using the provided path.
    *
    * Only `ico` and `png` are supported (based on activated feature flag).
@@ -117,9 +81,10 @@ export class Image extends Resource {
 }
 
 /**
- * Transforms image from various types into a type acceptable by Rust. Intended for internal use only.
+ * Transforms image from various types into a type acceptable by Rust.
  *
- * @ignore
+ * See [tauri::image::JsImage](https://docs.rs/tauri/2/tauri/image/enum.JsImage.html) for more information.
+ * Note the API signature is not stable and might change.
  */
 export function transformImage<T>(
   image: string | Image | Uint8Array | ArrayBuffer | number[] | null